+++ src/mongodb_session_manager/__init__.py
@@ -114,6 +114,6 @@ __all__.extend(
     ]
 )
 
-__version__ = "0.2.9"
+__version__ = "0.3.0"
 __author__ = "Iñaki Guinea Beristain"
 __author_email__ = "iguinea@gmail.com"